begin filter, change point to geo point in database

+++ backend/Application/UseCases/UpdateCrime/UpdateCrimeRequest.cs
@@ -9,6 +9,6 @@
         DateTime WantedPersonBirthDate,
         DateTime CrimeDate,
         string Location,
-        decimal PointLatitude,
-        decimal PointLongitude);
+        double PointLatitude,
+        double PointLongitude);
 }
